[ENH] add security file. Move purchase.type and purchase.method to purchase.py
