+++ pabi_procurement/models/purchase.py
@@ -112,3 +112,17 @@ class PurchaseOrderLine(models.Model):
         'purchase_requisition_line',
         'Purchase Requisition Line'
     )
+
+
+class PurchaseType(models.Model):
+    _name = 'purchase.type'
+    _description = 'PABI2 Purchase Type'
+
+    name = fields.Char(string='Purchase Type')
+
+
+class PurchaseMethod(models.Model):
+    _name = 'purchase.method'
+    _description = 'PABI2 Purchase Method'
+
+    name = fields.Char(string='Purchase Method')
